moved down IsCovered to GeometryService (avoid NetTopology usings)
